Add options to override default substituters and trusted-public-keys
This commit introduces a way to override the default substituters and trusted-public-keys copied from the reference machine.

A common use case is when the reference machine points to a local "proxy" binary cache (e.g., 127.0.0.1), which will not work on the target host.

Using `--option substituters <list>` or `--option extra-substituters <list>` does not solve this problem, since Nix now complains that the user is not part of trusted-users.

This commit also changes how `~/.config/nix/nix.conf` is created: it now stores a backup (with a `.orig` suffix) before patching the file, and reuses that backup on future invocations.

This avoids the situation where multiple runs of `nixos-anywhere` repeatedly append the same configuration over and over.